vcl: remove boost/signal2/signal.hpp from header
The most relevant signal member function appears to be connect(),
so let's create a wrapper function for that now, without the more
esoteric ordering features for now.

Move the signal member itself to a new pImpl.

The benefits are worth it: preprocessor input reduced by 2.8GB,
that's 9% of the total (excluding system headers which are not counted
because they don't generate dependencies).

Change-Id: I0aaeda51a5630a348bb12c81a83f67afbd508a14