tdf#92471: Improve color conversion to MSO highlighting
The simple color distance on RGB color space is not good enough.
It leads to a better result if we use the scheme of the primary
colors.
This method works well with MSO highlighting color palette, but
not neccessarily in general. In highlighting palette light and
dark variant of the same color type (e.g. blue and light blue) has
the exactly same scheme.

Change-Id: Ied08b4c388b8020326709d01d2de290afdd9d77b