tdf#89702 DOCX import: fix too large bullet character
Commit 3261a0a34d46e53a6841cee647943d9d6442fb61 (fdo#83465 RTF import:
handle font of numbering, 2014-09-25), changed the "get the me character
style of the current numbering's current level" member function to be
successfull even in case we're inside a DOCX run, not when we're inside
a DOCX paragraph, but outside runs.

While this is necessary for RTF, the side effect of this was that
unwanted run properties started to affect the above mentioned character
style in case of DOCX.

Fix the problem by enabling the "in paragraph and run" looking for RTF
only.

Change-Id: I610bfce6cec15b918fe547402360f5a894401f7e