tdf#92305: Revert "tdf#90150 : Page numbering in the status bar ..."
Unfortunately the original commit caused a regression.

The fix of tdf#90150 is also problematic from the UX point of view - with
that, the user is never sure what page he/she is looking at.

Should we do any changes there, then it should show the range of visible pages,
like "Pages 2-3 of 5." or "Pages 1-4 of 8." etc.

This reverts commit 6d122381825e272d99ae7f7ec7bda7c5c75a6124.

Change-Id: I371bb326da14b621a5116295ed1dc0de208f309d