tdf#93482 vcl rendercontext: introduce WindowImpl::mbDoubleBufferingRequested
The intention is that currently double-buffering is either enabled
globally or not. Double-buffering is known to be working in Writer, but
not in other applications, so it would be nice if double-buffering could
be also half-enabled: only in the applications where it's known to work.

For that, we need to differentiate between "we have a buffer" (supports
double buffering) and "we want to have a buffer if possible" (double
buffering requested).

Change-Id: If48d6dc0ddf5841497e78b856d803cc8abf23ac9