tdf#92379: svx: don't overwrite fo:background-color on ODF import
getSvxBrushItemFromSourceSet() is called once for each property of a
frame style that is set, and the "solid" fill-style is set by first
applying the BackColorRGB and then BackTransparency property.  So there
is an intermediate state that has fill-style NONE but a XFillColorItem
set - don't reset the color to white in that case!

Apparently writerfilter generally relies on the current reset-to-AUTO
implementation (loads of test failures), so for now limit the fix for
when we know that a frame style is imported from ODF.

(regression from 49635427f2798d9bd8641e1b3f70a09969115b64)

Change-Id: Ia987fe586ec819a1f3cd85acdbfb03c74ca0ec2c