Support modification write-back to other than registrymodifications.xcu
...to eventually support writing to dconf (see TODO).

Even when writing somewhere else, it may still be useful to read from the
current "user:" location, so a new convention was introduced to
CONFIGURATION_LAYERS types that support write-back:

- A leading "!" indicates that the layer is indeed used for write-back (probably
in addition to reading from it).  For backwards compatibility (when users use
own settings of CONFIGURATION_LAYERS, instead of depending on the value in the
shipped fundamental ini-file), no prefix on the "user:" is now interpreted the
same as a "!" prefix.

- A leading "*" indicates that the layer is not used for write-back (but only
for reading from it).

Change-Id: I399cc7bfe927db50586834f9630c184aaa2153f2