Revert "exclude unnamed Library/Executable from ide integration"
This reverts commit b51b69f2e0d79ee59da13d642c8554c8eb0cafc8.

We need to generate project files for unnamed projects as they seem to
correspond to libraries. If that causes problems for the msvc
integration (which seems strange) then we need to handle these problems
in the MSVC generator and not in the parser.

It also includes:

Revert "Clean trailing whitespace"

This reverts commit 9a9ae61fd83648e5e2282e84488cb4965fd58dfe.