sw: avoid asserts from RsidOnlyAutoFormats check
This happens with a document that has a header or footer that is not
displayed after loading, for example 1-page document with separate
left/right or first page header.

The SwTextNode::FileLoadedInitHints() is not called for these headers
because they are not visible so MakeFrm() is not called.

If they are not visible the RSID related flags don't matter.

Change-Id: Ie3bd1ce3bc0ac2ff1c429386149294e0c8fd3cb6