tdf#92505 vcl: fix substituting missing glyphs
Commit d0d2580dd5993a7ce5bdb017ad3b6cd748c86886 (Related bnc#822625:
Cache FontEntry with the original FontSelectPattern.) and the follow-up
a82f3c9ee936fe6e8cc1584d79e7b0914d01281b (Oops, this reference can't be
const., 2014-05-28) changed the font cache key, so that it's directly
what was the argument to ImplFontCache::GetFontEntry(), which leads to
missing glyphs in the bugdoc, e.g. the 4 check mark characters are
rendered as a check mark only in the DejaVu Sans case on my machine, the
other 3 cases result in placeholders.

Fix the problem by changing back the key to aFontSelData, which is
possibly modified in ImplFontCache::GetFontEntry() before used as a key.

Change-Id: I80eb4ba0f986a10100b6ae421d1346f235ce83b7