TableRef: fully refresh table column names also for partial range
Only one cell of a range was broadcasted per area listener if multiple
cells were affected. We don't know if there were more.

We could set up cell listeners instead, but that's not worth the hassle,
header cells don't change that often and if changed the header range to
refresh will not be overly large.

Change-Id: I915101b809fc1824803f10e696c52fb2185117f7