tdf#93139 Orca Screen Reader does not read bullet
To make Orca read bullet lists lets not suppress it
from the accessible text. This has a downside on Windows
when using NVDA which due to a hack in NVDA now reads 
the bullet twice. This hack is done because of this
bug so I think we should fix it.

Change-Id: Ibc11c1e0545f8cde4cd3d4a0d665ea47d645a61d
Reviewed-on: https://gerrit.libreoffice.org/18398
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Jacobo Aragunde Pérez <jaragunde@igalia.com>