slideshow: fix mysterious crash with the gtk3 vclplug
For some reason with gtk3 events are handled a bit differently, and in
particular after the PresenterSlideShowView::Resize() sets the
mbIsForcedPaintPending = true, with gtk2 we get a
notifySlideAnimationsEnded event and then a notifyViewChanged event that
calls PresenterSlideShowView::clear() to reset the flag,
but with gtk3 the flag isn't reset and then
PresenterSlideShowView::ForceRepaint() destroys the SlideView,
while there are still events in the EventQueue with pointers to it.

Since i'm evidently too dumb to tell what of this event handling is
working correctly and what is buggy, avoid the crash by checking
that the SlideView is still alive in the event handlers.

Change-Id: Ib88e61536c21e9787cef8a436341bfbd89914f4b