sfx2: add SfxViewShell::libreOfficeKitViewCallback()
This is similar to the existing LOK callback, the difference is that the
existing one assumes there is only one SfxViewShell instance at the same
time.

This newer callback is precisely per-view, so model notifications can
invoke all view callbacks, while view notifications can invoke only the
callback of the relevant view.

This is just the framework, all actual client code has to be still
ported over (and then the existing callback can be removed).

Change-Id: I3d8f27740c69fcf6ffbbdce12db2ea088321493d